Shrink ObservableMeasurementTableData
This introduces some new classes. Their naming is... not good.
