+++ qupath-gui-fx/src/main/java/qupath/lib/gui/measure/ObjectMeasurementListBuilder.java
@@ -1,0 +1,11 @@
+package qupath.lib.gui.measure;
+
+import qupath.lib.measurements.dynamic.MeasurementBuilder;
+
+import java.util.List;
+
+public interface ObjectMeasurementListBuilder {
+
+    List<MeasurementBuilder<?>> getMeasurements(PathObjectListWrapper wrapper);
+
+}
